+++ rust/candid/src/bindings/candid.rs
@@ -140,7 +140,7 @@ fn pp_fields(fs: &[Field], is_variant: bool) -> RcDoc {
     enclose_space("{", fields, "}")
 }
 
-fn pp_function(func: &Function) -> RcDoc {
+pub fn pp_function(func: &Function) -> RcDoc {
     let args = pp_args(&func.args);
     let rets = pp_args(&func.rets);
     let modes = pp_modes(&func.modes);